+++ docs/changelog.md
@@ -18,6 +18,9 @@ permalink: /changelog/
 * [Gecko](https://github.com/mozilla-mobile/android-components/blob/main/buildSrc/src/main/java/Gecko.kt)
 * [Configuration](https://github.com/mozilla-mobile/android-components/blob/main/.config.yml)
 
+* **browser-icons**:
+  * 🚒 Bug fixed [issue #10749](https://github.com/mozilla-mobile/android-components/issues/10749) - Add a 100ms timeout to returning a valid icon or default to a generated one to ensure other loadIcon() calls are not stalled.
+
 * **concept-toolbar**, **concept-engine**, **browser-engine-gecko**, **browser-state**, **feature-toolbar**, **browser-toolbar**,
   * 🌟️ The toolbar now supports two new methods: `expand` and `collapse` to immediately execute this actions if the toolbar is dynamic. `expand` is used as of now as a callback for when GeckoView needs the toolbar to be shown depending on tab content changes.